refactor kubeflow example
